+++ lib/nostr.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative 'nostr/errors'
 require_relative 'nostr/crypto'
 require_relative 'nostr/version'
 require_relative 'nostr/keygen'
@@ -14,6 +15,9 @@ require_relative 'nostr/event'
 require_relative 'nostr/events/encrypted_direct_message'
 require_relative 'nostr/client'
 require_relative 'nostr/user'
+require_relative 'nostr/key'
+require_relative 'nostr/private_key'
+require_relative 'nostr/public_key'
 
 # Encapsulates all the gem's logic
 module Nostr