+++ spec/nostr/errors/invalid_key_length_error_spec.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe Nostr::InvalidKeyLengthError do
+  describe '#initialize' do
+    let(:key_kind) { 'private' }
+    let(:error) { described_class.new(key_kind) }
+
+    it 'builds a useful error message' do
+      expect(error.message).to eq('Invalid private key length. It should have 64 characters.')
+    end
+  end
+end